requirements.txt
asgiref==3.11.1
dj-database-url==3.1.2
Django==4.2.29
django-cors-headers==4.9.0
django-filter==25.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.12.0
